requirements.txt
pillow>=8.0.0
aiohttp>=3.8.0